Test Coffeescript Posting

+++ db/migrate/20181102144609_create_pending_parks.rb
@@ -3,7 +3,7 @@ class CreatePendingParks < ActiveRecord::Migration[5.2]
     create_table :pending_parks do |t|
       t.string :uuid
       t.string :slug
-      t.integer :rvparky_id
+      t.bigint :rvparky_id
 
       t.timestamps
     end
